fix clear logs
