+++ rust/gui-client/src-tauri/src/client/logging.rs
@@ -5,7 +5,6 @@ use anyhow::{bail, Context, Result};
 use firezone_headless_client::known_dirs;
 use serde::Serialize;
 use std::{
-    ffi::OsStr,
     fs,
     io::{self, ErrorKind::NotFound},
     path::{Path, PathBuf},
@@ -112,78 +111,7 @@ pub(crate) async fn count_logs() -> Result<FileCount, String> {
 /// If we get an error while removing a file, we still try to remove all other
 /// files, then we return the most recent error.
 pub(crate) async fn clear_gui_logs() -> Result<()> {
-    let mut result = Ok(());
-
-    for log_path in log_paths()?.into_iter().map(|x| x.src) {
-        let mut dir = match tokio::fs::read_dir(log_path).await {
-            Ok(x) => x,
-            Err(error) => {
-                if matches!(error.kind(), NotFound) {
-                    // In smoke tests, the IPC service runs in debug mode, so it won't write any logs to disk. If the IPC service's log dir doesn't exist, we shouldn't crash, it's correct to simply not delete the non-existent files
-                    return Ok(());
-                }
-                // But any other error like permissions errors, should bubble.
-                return Err(error.into());
-            }
-        };
-        let mut paths = vec![];
-        while let Some(entry) = dir.next_entry().await? {
-            paths.push(entry.path());
-        }
-
-        let to_delete = choose_logs_to_delete(&paths);
-        for path in &to_delete {
-            if let Err(e) = tokio::fs::remove_file(path).await {
-                result = Err(e);
-            }
-        }
-    }
-
-    Ok(result?)
-}
-
-fn choose_logs_to_delete(paths: &[PathBuf]) -> Vec<&Path> {
-    let mut most_recent_stem = None;
-    for path in paths {
-        if path.extension() != Some(OsStr::new("log")) {
-            continue;
-        }
-        let Some(stem) = path.file_stem() else {
-            continue;
-        };
-        match most_recent_stem {
-            None => most_recent_stem = Some(stem),
-            Some(most_recent) if stem > most_recent => most_recent_stem = Some(stem),
-            Some(_) => {}
-        }
-    }
-    let Some(most_recent_stem) = most_recent_stem else {
-        tracing::warn!(
-            "Nothing to delete, should be impossible since both processes always write logs"
-        );
-        return vec![];
-    };
-    let Some(most_recent_stem) = most_recent_stem.to_str() else {
-        tracing::warn!("Most recent log file does not have a UTF-8 path");
-        return vec![];
-    };
-
-    paths
-        .iter()
-        .filter(|path| {
-            let Some(stem) = path.file_stem() else {
-                return false;
-            };
-            let Some(stem) = stem.to_str() else {
-                return false;
-            };
-            if !stem.starts_with("connlib.") {
-                return false;
-            }
-            stem < most_recent_stem
-        })
-        .map(|x| x.as_path())
-        .collect()
+    firezone_bin_shared::clear_logs(&known_dirs::logs().context("Can't compute GUI log dir")?).await
 }
 
 /// Pops up the "Save File" dialog
@@ -304,53 +232,12 @@ async fn count_one_dir(path: &Path) -> Result<FileCount> {
 fn log_paths() -> Result<Vec<LogPath>> {
     Ok(vec![
         LogPath {
-            src: firezone_headless_client::known_dirs::ipc_service_logs()
-                .context("Can't compute IPC service logs dir")?,
+            src: known_dirs::ipc_service_logs().context("Can't compute IPC service logs dir")?,
             dst: PathBuf::from("connlib"),
         },
         LogPath {
-            src: known_dirs::logs().context("Can't compute app log dir")?,
+            src: known_dirs::logs().context("Can't compute GUI log dir")?,
             dst: PathBuf::from("app"),
         },
     ])
-}
-
-#[cfg(test)]
-mod tests {
-    use std::path::Path;
-
-    #[test]
-    fn clear_logs_logic() {
-        // These are out of order just to make sure it works anyway
-        let paths: Vec<_> = [
-            "connlib.2024-08-05-19-41-46.jsonl",
-            "connlib.2024-08-05-19-41-46.log",
-            "connlib.2024-08-07-14-17-56.jsonl",
-            "connlib.2024-08-07-14-17-56.log",
-            "connlib.2024-08-06-14-21-13.jsonl",
-            "connlib.2024-08-06-14-21-13.log",
-            "connlib.2024-08-06-14-51-19.jsonl",
-            "connlib.2024-08-06-14-51-19.log",
-            "crash.2024-07-22-21-16-20.dmp",
-            "last_crash.dmp",
-        ]
-        .into_iter()
-        .map(|x| Path::new("/bogus").join(x))
-        .collect();
-        let to_delete = super::choose_logs_to_delete(&paths);
-        assert_eq!(
-            to_delete,
-            [
-                "/bogus/connlib.2024-08-05-19-41-46.jsonl",
-                "/bogus/connlib.2024-08-05-19-41-46.log",
-                "/bogus/connlib.2024-08-06-14-21-13.jsonl",
-                "/bogus/connlib.2024-08-06-14-21-13.log",
-                "/bogus/connlib.2024-08-06-14-51-19.jsonl",
-                "/bogus/connlib.2024-08-06-14-51-19.log",
-            ]
-            .into_iter()
-            .map(Path::new)
-            .collect::<Vec<_>>()
-        );
-    }
 }
